Added Settings View and SCRType Enum
A new Settings View was created to manage application settings. Additionally, an SCRType enum was introduced with three types: PWM, Voltage, and Current. The gradle file was modified to include dependencies for charts and German translations. Translations for SCRType and the new Settings View were also added to message properties files.