Refactor GPIO view layout and add getters in GPIOItem
The changes include a refactor in GPIOView.java to make use of HorizontalLayout for the gpioGrid instead of adding directly to the content. The GPIOItem.java class was modified to include getter methods for 'type' and 'name' attributes. Changes in gpio-view.xml reflect the updated layout structure in GPIOView.java.