Update theme styles, settings-view, and add system svg icon
The update includes changes to the PV-Heating css theme, enhancing the card element's hover and default styles. It also includes changes to the settings view and the addition of a new 'system' tab with its respective svg icon. Furthermore, a new system svg icon was also added as a new file in the theme's icon folder.